+++ R/checkClusterUniformity.R
@@ -28,8 +28,8 @@
 #'
 #' @importFrom withr local_options
 #'
-#' @importFrom zeallot `%<-%`
-#' @importFrom zeallot `%->%`
+#' @importFrom zeallot %<-%
+#' @importFrom zeallot %->%
 #'
 #' @export
 #'